solver: split llbop pkg
