+++ solver/solver/jobs.go
@@ -8,7 +8,9 @@ import (
 	"github.com/moby/buildkit/cache/instructioncache"
 	"github.com/moby/buildkit/client"
 	"github.com/moby/buildkit/session"
+	"github.com/moby/buildkit/solver"
 	"github.com/moby/buildkit/solver/pb"
+	"github.com/moby/buildkit/solver/reference"
 	"github.com/moby/buildkit/util/progress"
 	digest "github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
@@ -161,18 +163,18 @@ type job struct {
 
 type cacheRecord struct {
 	VertexSolver
-	index Index
-	ref   Reference
+	index solver.Index
+	ref   solver.Ref
 }
 
-func (j *job) load(def *pb.Definition, resolveOp ResolveOpFunc) (*Input, error) {
+func (j *job) load(def *pb.Definition, resolveOp ResolveOpFunc) (*solver.Input, error) {
 	j.l.mu.Lock()
 	defer j.l.mu.Unlock()
 
 	return j.loadInternal(def, resolveOp)
 }
 
-func (j *job) loadInternal(def *pb.Definition, resolveOp ResolveOpFunc) (*Input, error) {
+func (j *job) loadInternal(def *pb.Definition, resolveOp ResolveOpFunc) (*solver.Input, error) {
 	vtx, idx, err := loadLLB(def, func(dgst digest.Digest, pbOp *pb.Op, load func(digest.Digest) (interface{}, error)) (interface{}, error) {
 		if st, ok := j.l.actives[dgst]; ok {
 			if vtx, ok := st.jobs[j]; ok {
@@ -204,7 +206,7 @@ func (j *job) loadInternal(def *pb.Definition, resolveOp ResolveOpFunc) (*Input,
 			}
 			for i, input := range pbOp.Inputs {
 				if inputMetadata := def.Metadata[input.Digest]; inputMetadata.IgnoreCache {
-					k, err := s.CacheKey(ctx, Index(i))
+					k, err := s.CacheKey(ctx, solver.Index(i))
 					if err != nil {
 						return nil, err
 					}
@@ -225,7 +227,7 @@ func (j *job) loadInternal(def *pb.Definition, resolveOp ResolveOpFunc) (*Input,
 	if err != nil {
 		return nil, err
 	}
-	return &Input{Vertex: vtx.(*vertex), Index: idx}, nil
+	return &solver.Input{Vertex: vtx.(*vertex), Index: solver.Index(idx)}, nil
 }
 
 func (j *job) discard() {
@@ -253,12 +255,12 @@ func (j *job) getSolver(dgst digest.Digest) (VertexSolver, error) {
 	return st.solver, nil
 }
 
-func (j *job) getRef(ctx context.Context, v *vertex, index Index) (Reference, error) {
-	s, err := j.getSolver(v.Digest())
+func (j *job) getRef(ctx context.Context, cv client.Vertex, index solver.Index) (solver.Ref, error) {
+	s, err := j.getSolver(cv.Digest)
 	if err != nil {
 		return nil, err
 	}
-	ref, err := getRef(ctx, s, v, index, j.cache)
+	ref, err := getRef(ctx, s, cv, index, j.cache)
 	if err != nil {
 		return nil, err
 	}
@@ -266,15 +268,15 @@ func (j *job) getRef(ctx context.Context, v *vertex, index Index) (Reference, er
 	return ref, nil
 }
 
-func (j *job) keepCacheRef(s VertexSolver, index Index, ref Reference) {
-	immutable, ok := toImmutableRef(ref)
+func (j *job) keepCacheRef(s VertexSolver, index solver.Index, ref solver.Ref) {
+	immutable, ok := reference.ToImmutableRef(ref)
 	if ok {
 		j.cached[immutable.ID()] = &cacheRecord{s, index, ref}
 	}
 }
 
-func (j *job) cacheExporter(ref Reference) (CacheExporter, error) {
-	immutable, ok := toImmutableRef(ref)
+func (j *job) cacheExporter(ref solver.Ref) (CacheExporter, error) {
+	immutable, ok := reference.ToImmutableRef(ref)
 	if !ok {
 		return nil, errors.Errorf("invalid reference")
 	}
@@ -285,7 +287,7 @@ func (j *job) cacheExporter(ref Reference) (CacheExporter, error) {
 	return cr.Cache(cr.index, cr.ref), nil
 }
 
-func getRef(ctx context.Context, s VertexSolver, v *vertex, index Index, cache instructioncache.InstructionCache) (Reference, error) {
+func getRef(ctx context.Context, s VertexSolver, cv client.Vertex, index solver.Index, cache instructioncache.InstructionCache) (solver.Ref, error) {
 	k, err := s.CacheKey(ctx, index)
 	if err != nil {
 		return nil, err
@@ -295,8 +297,8 @@ func getRef(ctx context.Context, s VertexSolver, v *vertex, index Index, cache i
 		return nil, err
 	}
 	if ref != nil {
-		markCached(ctx, v.clientVertex)
-		return ref.(Reference), nil
+		markCached(ctx, cv)
+		return ref.(solver.Ref), nil
 	}
 
 	ev, err := s.OutputEvaluator(index)
@@ -316,8 +318,8 @@ func getRef(ctx context.Context, s VertexSolver, v *vertex, index Index, cache i
 				return nil, err
 			}
 			if ref != nil {
-				markCached(ctx, v.clientVertex)
-				return ref.(Reference), nil
+				markCached(ctx, cv)
+				return ref.(solver.Ref), nil
 			}
 			continue
 		}
