+++ solver/solver/load.go
@@ -3,6 +3,7 @@ package solver
 import (
 	"strings"
 
+	"github.com/moby/buildkit/solver"
 	"github.com/moby/buildkit/solver/pb"
 	"github.com/moby/buildkit/source"
 	digest "github.com/opencontainers/go-digest"
@@ -16,18 +17,18 @@ func newVertex(dgst digest.Digest, op *pb.Op, opMeta *pb.OpMetadata, load func(d
 		if err != nil {
 			return nil, err
 		}
-		vtx.inputs = append(vtx.inputs, &input{index: Index(in.Index), vertex: sub.(*vertex)})
+		vtx.inputs = append(vtx.inputs, &input{index: solver.Index(in.Index), vertex: sub.(*vertex)})
 	}
 	vtx.initClientVertex()
 	return vtx, nil
 }
 
-func toInternalVertex(v Vertex) *vertex {
+func toInternalVertex(v solver.Vertex) *vertex {
 	cache := make(map[digest.Digest]*vertex)
 	return loadInternalVertexHelper(v, cache)
 }
 
-func loadInternalVertexHelper(v Vertex, cache map[digest.Digest]*vertex) *vertex {
+func loadInternalVertexHelper(v solver.Vertex, cache map[digest.Digest]*vertex) *vertex {
 	if v, ok := cache[v.Digest()]; ok {
 		return v
 	}
@@ -43,7 +44,7 @@ func loadInternalVertexHelper(v Vertex, cache map[digest.Digest]*vertex) *vertex
 
 // loadLLB loads LLB.
 // fn is executed sequentially.
-func loadLLB(def *pb.Definition, fn func(digest.Digest, *pb.Op, func(digest.Digest) (interface{}, error)) (interface{}, error)) (interface{}, Index, error) {
+func loadLLB(def *pb.Definition, fn func(digest.Digest, *pb.Op, func(digest.Digest) (interface{}, error)) (interface{}, error)) (interface{}, pb.OutputIndex, error) {
 	if len(def.Def) == 0 {
 		return nil, 0, errors.New("invalid empty definition")
 	}
@@ -81,7 +82,7 @@ func loadLLB(def *pb.Definition, fn func(digest.Digest, *pb.Op, func(digest.Dige
 	}
 
 	v, err := rec(dgst)
-	return v, Index(lastOp.Inputs[0].Index), err
+	return v, lastOp.Inputs[0].Index, err
 }
 
 func llbOpName(op *pb.Op) string {
