+++ solver/solver/signal.go
